Collapse counts

Collapse counts

collapse_counts(
  counts_mat,
  meta_data,
  varnames,
  min_cells_per_group = 0,
  keep_n = FALSE,
  how = c("sum", "mean")[1]
)

Arguments

counts_mat

counts matrix where columns represent cells and rows represent features

meta_data

data.frame containing cell metadata

varnames

subset of `meta_data` column names

min_cells_per_group

minimum cells to keep collapsed group

keep_n

keep or drop the `N`` column containing the number of cells in each group. Default is `FALSE`

how

method of collapsing counts from groups. `sum` or `mean`

Examples

library(Seurat)
m <- matrix(sample.int(8, 100*500, replace=TRUE),nrow=100, ncol=500)
rownames(m) <- paste0("G", 1:100)
colnames(m) <- paste0("C", 1:500)
o <- CreateSeuratObject(m)
o$md1 <- sample(c("a", "b"), 500, replace=TRUE)
o$md2 <- sample(c("c", "d"), 500, replace=TRUE)
data_collapsed <- collapse_counts(
    o@assays$RNA@counts, o@meta.data, c('md1', 'md2')
)