presto: Fast differential expression

Scalable implementation of the Wilcoxon rank sum test and the area under the receiver operating characteristic curve (also known as auROC or AUC) statistic. Interfaces to dense and sparse matrices, as well as the genomics analysis frameworks 'Seurat' and 'SingleCellExperiment'.

Author

Maintainer: Kamil Slowikowski kslowikowski@gmail.com (ORCID) [contributor]

Authors:

Ilya Korsunsky ilya.korsunsky@gmail.com (ORCID)

Aparna Nathan

Nghia Millard (ORCID)

Soumya Raychaudhuri (ORCID)

Other contributors:

Austin Hartman (ORCID) [contributor]